fix: undo type undo-type-update
An update to the TS definition in https://github.com/jsonata-js/jsonata/pull/633
is causing issues in transformers-jsonata since we are using both 1.x and 2.x
versions in there and types do not align anymore. Reverting the change for now.
We overwrite the type of the NodeExpr in mappings anyway, and so this should be
fine for now.